+++ src/stories/components/OrderItemComponent.js
@@ -4,12 +4,16 @@ import { Text } from "native-base";
 
 class OrderItemComponent extends React.PureComponent {
   onTableClick = () => this.props.onTableClick(this.props.index);
+  onTableLongPress = () => this.props.onTableLongPress(this.props.index);
 
   render() {
     const { id, tableNo, isTakeAway } = this.props;
 
     return (
-      <TouchableOpacity onPress={this.onTableClick}>
+      <TouchableOpacity
+        onPress={this.onTableClick}
+        onLongPress={this.onTableLongPress}
+      >
         <View style={[styles.view, isTakeAway && styles.takeAwayView]}>
           <Text style={styles.orderText}>[ORDER-{id}]</Text>
           <Text style={styles.text}>Table No {tableNo}</Text>